requirements-dev.txt
asgiref==3.8.1
dj-database-url==2.1.0
dj-static==0.0.6
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
python-decouple==3.8
sqlparse==0.4.4
static3==0.7.0
typing_extensions==4.11.0
